Fix for MultiIndex attribute exception on Export_Record
On forms with no MultiIndex support, no enabled element was created with a default of false. The export would throw and object null exception when searching for the enabled attribute. Two changes were added: one to create the attribute by default and the other to handle the exception on configurations that already exist without the attribute.